requirements.txt
Django==4.2.5
python-docx
